Break out statistics schema repairs into a repairs module
A future PR will add repairs for events, states, etc
