+++ homeassistant/components/recorder/repairs/__init__.py
@@ -1,0 +1,1 @@
+"""Repairs for Recorder."""
